requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
einops==0.8.0
email_validator==2.2.0
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.15.4
fsspec==2024.6.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.7.1
httpcore==0.9.1
httptools==0.6.1
httpx==0.13.3
huggingface-hub==0.24.1
hyperframe==5.2.0
idna==2.10
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.24.4
packaging==24.1
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rfc3986==1.5.0
rich==13.7.1
safetensors==0.4.3
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
sympy==1.13.1
tbb==2021.13.0
timm==1.0.7
tokenizers==0.19.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.43.1
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.3
watchfiles==0.22.0
websockets==12.0
